Updated files and Create Person Object to show digger
