+++ phaser3-game/src/scripts/objects/Node.js
@@ -1,0 +1,28 @@
+//This not is part of Binary Search Tree object
+//It will be shown on the screen
+export class Node
+{ 
+    constructor(inputContent) 
+    {
+        
+        // What could be diplayed in the tree  
+        this.content  = inputContent;
+        // Javascript does not have pointers 
+        // to make a node just do the node  Node.leftChild = new Node(X)
+        this.parentNode = null;
+        this.leftChild = null;
+        this.rightChild = null; 
+    }
+    
+    
+    // remove(data) 
+    
+    
+    // Helper function 
+    // findMinNode() 
+    // getRootNode() 
+    // inorder(node) 
+    // preorder(node)                
+    // postorder(node) 
+    // search(node, data) 
+}
